Rename secureCreatives file and lint (#1203)
* Rename secureCreatives file and lint

* Updated package script for linting

* Use 'gulp run-tests' in package script for testing
